update file structure, remove unused extension, cleanup code
