+++ Demo/Demo/Enums/Intent.swift
@@ -1,0 +1,4 @@
+enum Intent: String {
+    case capture
+    case authorize
+}
