+++ Demo/Demo/Extensions/UIVIewController+Extensions.swift
@@ -13,13 +13,6 @@ extension UIViewController {
         return button
     }
 
-    func label(_ text: String) -> UILabel {
-        let label = UILabel()
-        label.font = UIFont.boldSystemFont(ofSize: 15)
-        label.text = text
-        return label
-    }
-
     func textField(
         placeholder: String? = nil,
         defaultValue: String? = nil,
